Move features to SDK
