+++ extensions/Bitwarden.Server.Sdk/src/Content/Features/IFeatureMetadata.cs
@@ -1,4 +1,4 @@
-namespace Bitwarden.Extensions.Hosting.Features;
+namespace Bitwarden.Server.Sdk.Features;
 
 internal interface IFeatureMetadata
 {
